+++ Changelog.md
@@ -18,6 +18,8 @@ rules on making a good Changelog.
 
 ### Changed
 
+- Sanitize rpaths (`--sanitize-rpaths`) is now the default behavior.
+  Opt out with new `--no-sanitize-rpaths` [#223](https://github.com/matthew-brett/delocate/pull/223)
 - Improved error message for when a MacOS target version is not met.
   [#211](https://github.com/matthew-brett/delocate/issues/211)
 - `delocate-fuse` is no longer available and will throw an error when invoked.